added parse url
