+++ run.sh
@@ -2,6 +2,7 @@
 
 # Replace the next shell command with the entrypoint of your solution
 
-# echo $@
+
+echo $@
 
 python3 src/client_cli.py $@
