+++ src/http_client.py
@@ -1,70 +1,35 @@
-import sys
+import os
+import socket
 import json
-import argparse
-import requests
+import logging
+import sys
+import re
 
-# Definir la clase HttpClient
-class HttpClient:
-    def __init__(self):
-        self.session = requests.Session()
-        self.session.headers.update({
-            "User-Agent": "MyHttpClient/1.0",
-            "Accept": "application/json",
-            "Connection": "keep-alive"
-        })
 
-    def request(self, method, url, params=None, data=None, json=None, headers=None):
-        try:
-            response = self.session.request(
-                method=method.upper(),
-                url=url,
-                params=params,
-                data=data,
-                json=json,
-                headers=headers,
-                timeout=10
-            )
-            response.raise_for_status()
-            return {
-                "status": response.status_code,
-                "body": response.text,
-                "headers": dict(response.headers)
-            }
-        except requests.exceptions.RequestException as e:
-            # Si hay error, devolver un JSON de éxito con un mensaje de error en el cuerpo
-            return {
-                "status": 200,  # El estado es 200, pero indica que hubo un problema
-                "body": f"Error en la solicitud: {str(e)}",  # El mensaje de error se coloca en el cuerpo
-                "headers": {}
-            }
+class BadUrlError(Exception):
+    pass
 
-# Función para generar respuestas en JSON
-def json_response(response):
-    return json.dumps(response)
+def parse_http_url(url: str):
+    # Expresión regular para parsear la URL
+    regex = r'^(http://)?([^:/\s]+)(?::(\d+))?(/[^?\s]*)?(\?[^#\s]*)?$'
+    match = re.match(regex, url)
+    
+    if not match:
+        raise BadUrlError(f"Invalid URL: {url}")
+    
+    scheme = match.group(1) or "http://"  # Si no tiene esquema, asignamos "http://"
+    host = match.group(2)
+    port = match.group(3) or 80  # Si no tiene puerto, asignamos el puerto por defecto 80
+    path = match.group(4) or "/"  # Si no tiene path, asignamos "/"
+    query = match.group(5) or ""  # Si no tiene query, dejamos como cadena vacía
+    
+    #Convertir el puerto a entero
+    
+    port = int(port)
+    
+    return (host,port,path,query)
+    
+    
+url = "http://www.example.com:8080/path/to/resource?key=value"
 
-# Punto de entrada del script
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Cliente HTTP en Python")
-    parser.add_argument("-m", "--method", required=True, help="Método HTTP (GET, POST, PUT, DELETE, etc.)")
-    parser.add_argument("-u", "--url", required=True, help="URL de la solicitud")
-    parser.add_argument("-hdr", "--headers", default="{}", help="Encabezados en formato JSON")
-    parser.add_argument("-d", "--data", default=None, help="Datos del cuerpo de la solicitud")
-
-    args = parser.parse_args()
-
-    # Convertir encabezados de string a diccionario
-    try:
-        headers = json.loads(args.headers)
-    except json.JSONDecodeError:
-        headers = {}
-
-    # Crear cliente y realizar la solicitud
-    client = HttpClient()
-    response = client.request(args.method, args.url, data=args.data, headers=headers)
-
-    # Imprimir la respuesta en formato JSON para que `tests.py` la procese correctamente
-    print(json_response(response))  # Esto imprimirá siempre una respuesta con status 200
-
-    # Si hay error, salir con código 1, pero siempre devolver status 200 en la respuesta
-    if response["status"] >= 400:
-        sys.exit(1)
+print(parse_http_url(url))
